requirements.txt
amqp==2.5.2
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
Babel==2.8.0
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.2
django-debug-toolbar==2.2
django-filter==2.2.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
flower==0.9.3
idna==2.8
importlib-metadata==1.5.0
isort==4.3.21
kombu==4.6.7
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
redis==3.4.1
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
tornado==5.1.1
typed-ast==1.4.1
urllib3==1.25.8
vine==1.3.0
whitenoise==5.0.1
wrapt==1.11.2
zipp==2.1.0
